Move to assets-subsystem; add initial unit test, start dependency injections
