--- jobs/asset-sender/src/main/java/com/paladincloud/assetsender/Main.java
@@ -1,8 +1,0 @@
-package com.paladincloud.assetsender;
-
-public class Main {
-
-    public static void main(String[] args) {
-        System.exit(new AssetSenderJob().run(args));
-    }
-}
