+++ jobs/assets-subsystem/asset-sender/src/main/java/com/paladincloud/assetsender/AssetSenderJob.java
@@ -1,12 +1,13 @@
 package com.paladincloud.assetsender;
 
 import com.paladincloud.common.assets.Assets;
+import com.paladincloud.common.config.AssetTypes;
 import com.paladincloud.common.config.ConfigConstants;
 import com.paladincloud.common.config.ConfigService;
-import com.paladincloud.common.config.Types;
 import com.paladincloud.common.jobs.JobExecutor;
 import java.util.Collections;
 import java.util.List;
+import javax.inject.Inject;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
@@ -17,6 +18,16 @@ public class AssetSenderJob extends JobExecutor {
     public static final String TENANT_ID = "tenant-id";
     private static final Logger LOGGER = LogManager.getLogger(AssetSenderJob.class);
 
+    private final AssetTypes assetTypes;
+    private final Assets assets;
+
+    @Inject
+    AssetSenderJob(AssetTypes assetTypes, Assets assets) {
+        this.assetTypes = assetTypes;
+        this.assets = assets;
+    }
+
+
     @Override
     protected int execute() {
 
@@ -25,8 +36,8 @@ public class AssetSenderJob extends JobExecutor {
             params.get(S3_PATH), params.get(TENANT_ID));
         ConfigService.setProperties("batch.",
             Collections.singletonMap("s3.data", params.get(S3_PATH)));
-        Types.setupIndexAndTypes(params.get(DATA_SOURCE));
-        Assets.upload(params.get(DATA_SOURCE));
+        assetTypes.setupIndexAndTypes(params.get(DATA_SOURCE));
+        assets.upload(params.get(DATA_SOURCE));
 
         return 0;
     }
