+++ jobs/assets-subsystem/asset-sender/src/main/java/com/paladincloud/assetsender/Main.java
@@ -1,0 +1,11 @@
+package com.paladincloud.assetsender;
+
+import com.paladincloud.common.DaggerServerComponent;
+
+public class Main {
+
+    public static void main(String[] args) {
+        var componentResolver = DaggerServerComponent.create();
+        System.exit(componentResolver.buildAssetSenderJob().run(args));
+    }
+}
