+++ jobs/assets-subsystem/asset-sender/src/main/java/com/paladincloud/common/ServerComponent.java
@@ -1,0 +1,12 @@
+package com.paladincloud.common;
+
+import com.paladincloud.assetsender.AssetSenderJob;
+import dagger.Component;
+import javax.inject.Singleton;
+
+@Singleton
+@Component(modules = {ApplicationModule.class})
+public interface ServerComponent {
+
+    AssetSenderJob buildAssetSenderJob();
+}
