+++ jobs/assets-subsystem/asset-sender/src/main/java/com/paladincloud/common/assets/AssetGroups.java
