+++ jobs/assets-subsystem/asset-sender/src/main/java/com/paladincloud/common/errors/JobException.java
