+++ jobs/assets-subsystem/asset-sender/src/test/java/com/paladincloud/commons/fakes/TestComponent.java
@@ -1,0 +1,10 @@
+package com.paladincloud.commons.fakes;
+
+import com.paladincloud.common.ServerComponent;
+import dagger.Component;
+import javax.inject.Singleton;
+
+@Singleton
+@Component(modules = { TestModule.class })
+public interface TestComponent extends ServerComponent {
+}
