+++ jobs/assets-subsystem/asset-sender/src/test/java/com/paladincloud/commons/fakes/TestModule.java
@@ -1,0 +1,10 @@
+package com.paladincloud.commons.fakes;
+
+import dagger.Module;
+import dagger.Provides;
+import javax.inject.Singleton;
+
+@Module
+public class TestModule {
+
+}
